+++ server/src/dependencies.php
@@ -9,7 +9,10 @@ $container['db'] = function($c) {
 	return $pdo;
 };
 $container['UserManager'] = function($c) {
-	return new \LookupServer\UserManager($c->db, $c->EmailValidator);
+	return new \LookupServer\UserManager($c->db, $c->EmailValidator, $c->WebsiteValidator, $c->SignatureHandler);
+};
+$container['SignatureHandler'] = function($c) {
+	return new \LookupServer\SignatureHandler();
 };
 $container['EmailValidator'] = function($c) {
 	return new \LookupServer\Validator\Email(
@@ -19,6 +22,10 @@ $container['EmailValidator'] = function($c) {
 		$c->settings['emailfrom']
 	);
 };
+$container['WebsiteValidator'] = function($c) {
+	return new \LookupServer\Validator\Website($c->SignatureHandler);
+};
+
 $container['Status'] = function($c) {
 	return new \LookupServer\Status();
 };
